# Reduce lexer memory allocs