@diegommm 2025-06-30
> I just realized: why ring buffer needed at all?

I tried avoiding it but couldn't. My first attempt was to make the iterator run one state function and get it's resulting token. But then I found out that one state function call can produce more than one token.

In that scenario, then the tokens should be put in a slice, and then the slice consumed one at a time by the iterator until it was depleted, then clear the slice and call again the next state function. I did this too but the code started to look a ugly and hard to read. So I reused some code I wrote for an open source company for a ring buffer (adapted to make it smaller) and the changes to the lexer itself became less complex and the code looks cleaner.

As I hadn't touched the parser itself yet and I saw that it also needs to lookahead, I also thought that I would have to do something similar there so I thought that I would need to reuse the ring buffer. But fortunately I realized that the parser only needs to lookahead a single token at most (in the case of `not`) so it was far easier to just stash that single token and swap it if we wanted to rollback.

But if we can make it even simpler and remove the ring buffer then it's alright, I just saw it fit and keep the code clean and lean at that time.